requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
comtypes==1.1.10
cryptography==35.0.0
cssselect==1.1.0
cssutils==2.3.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
docopt==0.6.2
email-to==0.1.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
Js2Py==0.71
jwcrypto==1.0
lxml==4.6.4
Markdown==3.3.4
MouseInfo==0.1.3
oauth2client==4.1.3
packaging==21.0
Pillow==8.4.0
premailer==3.10.0
protobuf==3.19.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
pycryptodome==3.11.0
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.0.4
pyperclip==1.8.2
PyPrind==2.11.3
Pyrebase4==4.5.0
PyRect==0.1.4
PyScreeze==0.1.28
pySmartDL==1.3.4
python-jwt==3.3.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
secure-smtplib==0.1.1
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
wikipedia==1.4.0
wrapt==1.13.3
yagmail==0.14.260
